[REL] v3.2 - Update Odoo version to 16.0. - Dark mode for options page. - Adopt Odoo dark mode for navigation button.
