+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v3.2 - 2022-12-27
+### Added
+- Update Odoo version to 16.0.
+- Dark mode for options page.
+- Adopt Odoo dark mode for navigation button.
+
+
 ## v3.1 - 2022-10-04
 ### Added
 - Dark theme.
