+++ package.json
@@ -27,8 +27,8 @@
   },
   "homepage": "https://github.com/Droggol/odoo-utility#readme",
   "devDependencies": {
-    "bootstrap": "^5.1.3",
-    "node-sass": "^7.0.1",
-    "nodemon": "^2.0.15"
+    "bootstrap": "^5.3.0-alpha1",
+    "node-sass": "^8.0.0",
+    "nodemon": "^2.0.20"
   }
 }
