+++ public/options.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <title>Odoo Utility Options</title>
-        <link rel="stylesheet" href="/src/lib/icons-1.8.0/bootstrap-icons.css"/>
+        <link rel="stylesheet" href="/src/lib/icons-1.10.2/bootstrap-icons.css"/>
         <link type="text/css" href="/src/scss/style.css" rel="stylesheet"/>
         <script src="/src/js/options.js" type="module"></script>
     </head>
