+++ public/popup.html
@@ -2,9 +2,10 @@
 <html>
     <head>
         <title>Odoo Utility</title>
-        <link rel="stylesheet" href="/src/lib/icons-1.8.0/bootstrap-icons.css"/>
+        <link rel="stylesheet" href="/src/lib/icons-1.10.2/bootstrap-icons.css"/>
         <link rel="stylesheet" href="/src/scss/style.css"/>
         <script src="/src/lib/bootstrap.bundle.js"></script>
+        <script src="/src/js/common.js" type="module"></script>
         <script src="/src/js/popup.js" type="module"></script>
     </head>
     <body>
@@ -22,7 +23,7 @@
                         </div>
                     </div>
                     <div class="col-auto">
-                        <a class="btn btn-link ou-open-options-page" title="Open Options"><i class="bi-gear"></i></a>
+                        <a href="#" class="link-primary ou-open-options-page" title="Open Options"><i class="bi-gear"></i></a>
                     </div>
                 </div>
             </div>
