+++ public/src/lib/icons-1.10.2/LICENSE.md
